Bonus function & ubuntu compiler compatibility - Added the bonus option -c so that it colors the number that is being moved with every operation - Changed several files in my ft_printf and my libft so that it runs smoothly when compiled in an ubuntu machine - Corrected the leak when executables are run with no arguments or with wrong instructions

+++ shared_files/ft_list_functions.c
@@ -6,7 +6,7 @@
 /*   By: dsaripap <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/02/24 14:54:47 by dsaripap      #+#    #+#                 */
-/*   Updated: 2020/04/29 18:57:06 by dominique     ########   odam.nl         */
+/*   Updated: 2020/05/29 10:51:49 by dominique     ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,10 +41,10 @@ void				ft_stack_addend(t_stack_list **lst, t_stack_list *new)
 	new->prev = temp;
 }
 
-int					ft_stack_len(t_stack_list *lst)
+size_t				ft_stack_len(t_stack_list *lst)
 {
 	t_stack_list	*temp;
-	int				len;
+	size_t			len;
 
 	temp = lst;
 	len = 0;
